% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_webdriver.R
\name{browser_path}
\alias{browser_path}
\title{Return browser (webdriver) location path}
\usage{
browser_path()
}
\value{
path as character
}
\description{
After installing webdriver using \code{\link{install_browser}},
you can check its location path by running this function.
}
\examples{

\dontrun{

browser_paths()
}
}
\author{
salim khalil
}
