% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkExtractor.R
\name{Drv_fetchpage}
\alias{Drv_fetchpage}
\title{Fetch page using web driver/Session}
\usage{
Drv_fetchpage(url, browser)
}
\arguments{
\item{url}{character, web page URL to retreive}

\item{browser}{Object returned by \code{\link{run_browser}}}
}
\value{
return a list of three elements, the first is a list containing the web page details (url, encoding-type, content-type, content ... etc), the second is a character-vector containing the list of retreived internal urls and the third is a vetcor of external Urls.
}
\description{
Fetch page using web driver/Session
}
\author{
salim khalil
}
