% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getencoding.R
\name{Getencoding}
\alias{Getencoding}
\title{Getencoding}
\usage{
Getencoding(url)
}
\arguments{
\item{url}{character, url to}
}
\value{
return the encoding charset as character
}
\description{
A function that parse a web page content and retreive it's encoding charset based on content and HTTP header
}
\details{
xxx
}
\author{
salim khalil
}
\seealso{
\code{other function}
}

