% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsimHash.R
\name{getsimHash}
\alias{getsimHash}
\title{Calculate SimHash fingerprint in R}
\usage{
getsimHash(string, hashbits)
}
\arguments{
\item{string}{character, the content to hash.}

\item{hashbits}{numeric, specify the hash bits 64 or 128}
}
\value{
return the simhash as a nmeric value
}
\description{
A function that take a _charachter_ as input, and generate it's simhash.
}
\details{
This funcion call an external java class
}
\examples{

text<-"R is a free software environment for statistical computing and graphics.
      It compiles and runs on a wide variety of UNIX platforms, Windows and MacOS"
fingerprint<-getsimHash(text,64)


}
\author{
salim khalil
}
