% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Linkparameters.R
\name{Linkparameters}
\alias{Linkparameters}
\title{Link parameters}
\usage{
Linkparameters(URL)
}
\arguments{
\item{URL}{character, the URL to extract}
}
\value{
return the URL paremeters=values
}
\description{
A function that take a URL _charachter_ as input, and extract the parameters and values from this URL .
}
\details{
This function extract the link parameters and values (Up to 10 parameters)
}
\author{
salim khalil
}

