% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListProjects.R
\name{ListProjects}
\alias{ListProjects}
\title{ListProjects}
\usage{
ListProjects(DIR)
}
\arguments{
\item{DIR}{character By default it's your local R workspace, if you set a custom folder for your crawling project then user DIR param to access this folder.}
}
\value{
\code{ListProjects}, a character vector.
}
\description{
List all crawling project in your R local directory, or in a custom directory
}
\examples{
\dontrun{
ListProjects()
}
}
\author{
salim khalil
}
