% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadHTMLFiles.R
\name{LoadHTMLFiles}
\alias{LoadHTMLFiles}
\title{LoadHTMLFiles}
\usage{
LoadHTMLFiles(ProjectName, type = "vector", max, DIR)
}
\arguments{
\item{ProjectName}{character, the name of the folder holding collected HTML files, use \code{ListProjects} fnuction to see all projects.}

\item{type}{character, the type of returned variable, either vector or list.}

\item{max}{Integer, maximum number of files to load.}
}
\value{
\code{LoadHTMLFiles}, a character vector or a list;
}
\description{
LoadHTMLFiles
}
\examples{
\dontrun{
ListProjects()
#show all crawling project folders stored in your local R wokspace folder
DataHTML<-LoadHTMLFiles("glofile.com-301010")
#Load all HTML files in DataHTML vector
DataHTML2<-LoadHTMLFiles("glofile.com-301010",max = 10, type = "list")
#Load only 10 first HTMl files in DataHTML2 list
}
}
\author{
salim khalil
}
