% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_webdriver.R
\name{stop_browser}
\alias{stop_browser}
\title{Stop web driver process and Remove its Object}
\usage{
stop_browser(browser)
}
\arguments{
\item{browser}{the web driver object created by \code{\link{run_browser}}}
}
\value{
A list of \code{process}, the \code{callr::process} object, and
  \code{port}, the local port where phantom is running.
}
\description{
At the end of All your operations with the web river, you should stop its process
and remove the driver R object else you may have troubles restarting R normaly.
Throws and error if webdriver phantomjs cannot be found, or cannot be started.
It works with a timeout of five seconds.
}
\examples{
\dontrun{

#Start the browser
br<-run_browser()

#kill the browser process
stop_browser(br)
#remove the object reference
rm(br)

 }


}
