% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contentscraper.R
\name{contentscraper}
\alias{contentscraper}
\title{contentscraper}
\usage{
contentscraper(webpage, patterns, patnames, excludepat, astext = TRUE, encod)
}
\arguments{
\item{webpage}{character, a web page as text}

\item{patterns}{character vector, one or more XPath patterns to extract from the web page}

\item{patnames}{character vector, names of patterns}
}
\value{
return a named list of extracted content
}
\description{
A function for extracting content matching a specific XPath patterns from a web page, the function takes a web page text as _character_ and a set of named patterns then it return a list contining content matching patterns
}
\details{
contentscraper(x ,c("//head/title","//body/div/article"),c("title", "article"))
}
\author{
salim khalil
}

