% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getencoding.R
\name{Getencoding}
\alias{Getencoding}
\title{Getencoding}
\usage{
Getencoding(url)
}
\arguments{
\item{url}{character, the web page url.}
}
\value{
return the encoding charset as character
}
\description{
This function  retreives the encoding charset of web page based on HTML tags and HTTP header
}
\author{
salim khalil
}
