% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getsimHash.R
\name{getsimHash}
\alias{getsimHash}
\title{getsimHash}
\usage{
getsimHash(string, hashbits)
}
\arguments{
\item{string}{character, the content to hash.}

\item{hashbits}{numeric, specify the hash bits 64 or 128}
}
\value{
return the simhash as a nmeric value
}
\description{
A function that take a _charachter_ as input, and generate it's simhash.
}
\details{
This funcion call an external java class
}
\author{
salim khalil
}

