% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinkNormalization.R
\name{LinkNormalization}
\alias{LinkNormalization}
\title{Link Normalization}
\usage{
LinkNormalization(links, current)
}
\arguments{
\item{current}{character, The URL of the current page source of the link.}

\item{link}{character, the URL to Normalize.}
}
\value{
return the simhash as a nmeric value
}
\description{
A function that take a URL _charachter_ as input, and transforms it into a canonical form.
}
\details{
This funcion call an external java class
}
\author{
salim khalil
}

