% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkNormalization.R
\name{LinkNormalization}
\alias{LinkNormalization}
\title{Link Normalization}
\usage{
LinkNormalization(links, current)
}
\arguments{
\item{links}{character, the URL to Normalize.}

\item{current}{character, The URL of the current page source of the link.}
}
\value{
return the simhash as a nmeric value
}
\description{
A function that take a URL _charachter_ as input, and transforms it into a canonical form.
}
\details{
This funcion call an external java class
}
\examples{

# Normalize a set of links

links<-c("http://www.twitter.com/share?url=http://glofile.com/page.html",
         "/finance/banks/page-2017.html",
         "./section/subscription.php",
         "//section/",
         "www.glofile.com/home/",
         "glofile.com/sport/foot/page.html",
         "sub.glofile.com/index.php",
         "http://glofile.com/page.html#1"
                   )

links<-LinkNormalization(links,"http://glofile.com" )


}
\author{
salim khalil
}
