% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linkparameters.R
\name{Linkparameters}
\alias{Linkparameters}
\title{Get the list of parameters and values from an URL}
\usage{
Linkparameters(URL)
}
\arguments{
\item{URL}{character, the URL to extract}
}
\value{
return the URL paremeters=values
}
\description{
A function that take a URL _charachter_ as input, and extract the parameters and values from this URL .
}
\details{
This function extract the link parameters and values (Up to 10 parameters)
}
\examples{

Linkparameters("http://www.glogile.com/index.php?name=jake&age=23&template=2&filter=true")
# Extract all URL parameters with values as vector

}
\author{
salim khalil
}
