% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcrawlerp.R
\name{Rcrawler}
\alias{Rcrawler}
\title{Rcrawler}
\usage{
Rcrawler(Website, no_cores, nbcon, MaxDepth, DIR, RequestsDelay = 0,
  duplicatedetect = FALSE, Obeyrobots = FALSE, IndexErrPages, Useragent,
  Timeout = 5, URLlenlimit = 255, urlExtfilter, urlregexfilter,
  ignoreUrlParams, statslinks = FALSE, Encod, patterns, excludepattern,
  Backup = FALSE)
}
\arguments{
\item{Website}{character, the root URL of the web site to crawl}

\item{no_cores}{integer, specify the number of clusters for parallel crawling, by default is the numbers of available cores – 1}

\item{MaxDepth}{integer, repsent the max deph level for the crawler, this is not the file depth in a directory structure, but 1+ number of links between this document and root document, default to 10}

\item{DIR}{character, correspond to the path of the local repository the crawler will use to store crawled data ex, "C:/collection" , by default R working directory}
}
\value{
return the simhash as a nmeric value
}
\description{
A function that take a _charachter_ as input, and generate it's simhash.
}
\details{
This funcion call an external java class
for arabic use Sys.setlocale("LC_CTYPE","Arabic_Saudi Arabia.1256") and set encoding of the web page
}
\author{
salim khalil
}

